requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.54
click==6.7
docutils==0.13.1
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.2
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
python-dateutil==2.6.0
s3transfer==0.1.10
six==1.10.0
Werkzeug==0.12.2
